refactor: 결과 페이지 데이터 hook으로 추상화
